Recentered Camera anchor to vertex and added commandline nav tool. Also updated README.md

+++ Graphing_Terminal/base.h
@@ -28,7 +28,7 @@ typedef struct ScreenVectors{
 typedef struct Cameras
 {
 	Vector position;
-	float scale;
+	double scale;
 } Camera;
